requirements.txt
pyyaml
# nltk
# numpy
